requirementsProduction.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
Faker==15.0.0
firebase-admin==5.3.0
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.1
grpcio-status==1.48.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
include==0.2.2
msgpack==1.0.4
Pillow==9.3.0
proto-plus==1.22.1
protobuf==4.21.6
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
routers==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toposort==1.7
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
